+++ app/controllers/api/v2/contacts_controller.rb
@@ -2,6 +2,7 @@ class Api::V2::ContactsController < Api::V2Controller
   def index
     authorize_index
     load_contacts
+
     render json: @contacts.preload_valid_associations(include_associations)
       .preload(attributes_to_preload),
            meta: meta_hash(@contacts),